Updated versionLucasonssss stated that the opening crawl was inspired by the opening crawls used at the beginning of each episode of the original Flash Gordon and Buck Rogers film serials, which were the inspiration for Lucas to write much of the Star Wars saga.[3]
In a 2005 interview, George Lucas described how the final phrasing of the text for A New Hope came about. "The crawl is such a hard thing because you have to be careful that you're not using too many words that people don't understand. It's like a poem. I showed the very first crawl to a bunch of friends of mine in the 1970s. It went on for six paragraphs with four sentences each. Brian De Palma was there...". De Palma helped to edit the text into the form used in the film.[4]